Start deprecating use of pkg_resources
the use of pkg_resources breaks pytest>=5.0
We can circumvent the issue by moving to importlib_resources, but up
until last week this did not support the required functionality in a
clean manner.

The change in
https://gitlab.com/python-devs/importlib_resources/-/merge_requests/76
which is due to be included in Python 3.9 and the as of yet unreleased
importlib_resources package 1.1.0 should allow us to clean things up
again.